+++ src/components/ui/Input/Default/styles.ts
@@ -1,0 +1,9 @@
+import { css } from '@emotion/react';
+
+import { colors } from '@/styles/variants/theme';
+
+export const inputStyle = css({
+  border: `1px solid ${colors.newGray[200]}`,
+  borderRadius: '0.3rem',
+  width: '100%',
+});
