+++ src/pages/ProductPage/components/ProductForm/SelectNumber/styles.ts
@@ -1,7 +1,5 @@
 import { css } from '@emotion/react';
 
-import { colors } from '@/styles/variants/theme';
-
 export const cardStyle = css({
   padding: '1rem',
 });
@@ -11,10 +9,4 @@ export const buttonStyle = css({
   justifyContent: 'center',
   alignItems: 'center',
   width: '4rem',
-});
-
-export const inputStyle = css({
-  border: `1px solid ${colors.gray[200]}`,
-  borderRadius: '0.3rem',
-  width: '100%',
 });
